requirements.txt
# Core
python-telegram-bot==20.7
python-dotenv==1.0.0

# Google Sheets
gspread==5.12.0
oauth2client==4.1.3

# Data Processing
pandas==1.5.3
openpyxl==3.1.2

# Google Gemini AI
google-generativeai>=0.8.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Development (optional)
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Web Server
flask==2.3.3
gunicorn==21.2.0
